requirements.txt

pandas
scikit-learn
pytest
flake8
joblibs
